(PC-12002) feat(identityCheck) Enable idCheckAddressAutocompletion for SetAddress
